+++ src/utils/validating/checkers.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   validator.c                                        :+:      :+:    :+:   */
+/*   checkers.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ysaroyan <ysaroyan@student.42.fr>          +#+  +:+       +#+        */
+/*   By: vapetros <vapetros@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/05 19:56:36 by ysaroyan          #+#    #+#             */
-/*   Updated: 2025/04/13 16:46:08 by ysaroyan         ###   ########.fr       */
+/*   Updated: 2025/04/13 20:25:16 by vapetros         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,8 +30,8 @@ bool	is_identifier(char *str)
 		|| !ft_strcmp(str, CYLINDER) || !ft_strcmp(str, PLANE));
 }
 
-bool	is_instruction_in_range(char **inst, double min,
-	double max, int size)
+bool	is_instruction_in_range(char **inst, float min,
+	float max, int size)
 {
 	int	i;
